refactored feature login and wrapped it into feature auth
